requirements.txt
asgiref==3.8.1
Django==5.2.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.1
sqlparse==0.5.3
typing_extensions==4.14.0
